Mnist fix (#392)
* fixed mnist

* fixed mnist
